Add updating unit knowledge process
