+++ app/interactors/user_answers/create.rb
@@ -1,20 +1,7 @@
 module UserAnswers
   class Create
-    include Interactor
+    include Interactor::Organizer
 
-    delegate :user_answer, to: :context
-    delegate :test_question, to: :user_answer
-    delegate :question, to: :test_question
-
-    def call
-      user_answer.truthy = check_answer
-      user_answer.save! || context.fail!(error: user_answer.errors)
-    end
-
-    private
-
-    def check_answer
-      user_answer.answer.downcase.strip.eql?(question.right_answer.downcase)
-    end
+    organize Save, UnitKnowledges::Update
   end
 end
